third_party/blink/web_tests/platform/generic/fast/dom/Window/window-lookup-precedence-expected.txt

Tests lookup precedence for properties on Window objects in the presence of a child frame, and that IDL properties defined in the prototype (window.location, window.focus, etc) have higher lookup precedence than named document elements (such as input elements).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".




PASS 'function alert() { [native code] }' is 'function alert() { [native code] }'
PASS 'function alert() { [native code] }' is 'function alert() { [native code] }'
PASS 'function atob() { [native code] }' is 'function atob() { [native code] }'
PASS 'function atob() { [native code] }' is 'function atob() { [native code] }'
PASS 'function blur() { [native code] }' is 'function blur() { [native code] }'
PASS 'function blur() { [native code] }' is 'function blur() { [native code] }'
PASS 'function btoa() { [native code] }' is 'function btoa() { [native code] }'
PASS 'function btoa() { [native code] }' is 'function btoa() { [native code] }'
PASS 'function captureEvents() { [native code] }' is 'function captureEvents() { [native code] }'
PASS 'function captureEvents() { [native code] }' is 'function captureEvents() { [native code] }'
PASS 'function clearInterval() { [native code] }' is 'function clearInterval() { [native code] }'
PASS 'function clearInterval() { [native code] }' is 'function clearInterval() { [native code] }'
PASS 'function clearTimeout() { [native code] }' is 'function clearTimeout() { [native code] }'
PASS 'function clearTimeout() { [native code] }' is 'function clearTimeout() { [native code] }'
PASS 'function close() { [native code] }' is 'function close() { [native code] }'
PASS 'function close() { [native code] }' is 'function close() { [native code] }'
PASS 'function confirm() { [native code] }' is 'function confirm() { [native code] }'
PASS 'function confirm() { [native code] }' is 'function confirm() { [native code] }'
PASS 'function find() { [native code] }' is 'function find() { [native code] }'
PASS 'function find() { [native code] }' is 'function find() { [native code] }'
PASS 'function focus() { [native code] }' is 'function focus() { [native code] }'
PASS 'function focus() { [native code] }' is 'function focus() { [native code] }'
PASS 'function getComputedStyle() { [native code] }' is 'function getComputedStyle() { [native code] }'
PASS 'function getComputedStyle() { [native code] }' is 'function getComputedStyle() { [native code] }'
PASS 'function getSelection() { [native code] }' is 'function getSelection() { [native code] }'
PASS 'function getSelection() { [native code] }' is 'function getSelection() { [native code] }'
PASS 'function moveBy() { [native code] }' is 'function moveBy() { [native code] }'
PASS 'function moveBy() { [native code] }' is 'function moveBy() { [native code] }'
PASS 'function moveTo() { [native code] }' is 'function moveTo() { [native code] }'
PASS 'function moveTo() { [native code] }' is 'function moveTo() { [native code] }'
PASS 'function open() { [native code] }' is 'function open() { [native code] }'
PASS 'function open() { [native code] }' is 'function open() { [native code] }'
PASS 'function print() { [native code] }' is 'function print() { [native code] }'
PASS 'function print() { [native code] }' is 'function print() { [native code] }'
PASS 'function prompt() { [native code] }' is 'function prompt() { [native code] }'
PASS 'function prompt() { [native code] }' is 'function prompt() { [native code] }'
PASS 'function releaseEvents() { [native code] }' is 'function releaseEvents() { [native code] }'
PASS 'function releaseEvents() { [native code] }' is 'function releaseEvents() { [native code] }'
PASS 'function resizeBy() { [native code] }' is 'function resizeBy() { [native code] }'
PASS 'function resizeBy() { [native code] }' is 'function resizeBy() { [native code] }'
PASS 'function resizeTo() { [native code] }' is 'function resizeTo() { [native code] }'
PASS 'function resizeTo() { [native code] }' is 'function resizeTo() { [native code] }'
PASS 'function scroll() { [native code] }' is 'function scroll() { [native code] }'
PASS 'function scroll() { [native code] }' is 'function scroll() { [native code] }'
PASS 'function scrollBy() { [native code] }' is 'function scrollBy() { [native code] }'
PASS 'function scrollBy() { [native code] }' is 'function scrollBy() { [native code] }'
PASS 'function scrollTo() { [native code] }' is 'function scrollTo() { [native code] }'
PASS 'function scrollTo() { [native code] }' is 'function scrollTo() { [native code] }'
PASS 'function setInterval() { [native code] }' is 'function setInterval() { [native code] }'
PASS 'function setInterval() { [native code] }' is 'function setInterval() { [native code] }'
PASS 'function setTimeout() { [native code] }' is 'function setTimeout() { [native code] }'
PASS 'function setTimeout() { [native code] }' is 'function setTimeout() { [native code] }'
PASS 'function stop() { [native code] }' is 'function stop() { [native code] }'
PASS 'function stop() { [native code] }' is 'function stop() { [native code] }'
PASS 'function toString() { [native code] }' is 'function toString() { [native code] }'
PASS 'function toString() { [native code] }' is 'function toString() { [native code] }'
PASS 'function Attr() { [native code] }' is 'function Attr() { [native code] }'
PASS 'function Attr() { [native code] }' is 'function Attr() { [native code] }'
PASS 'function CDATASection() { [native code] }' is 'function CDATASection() { [native code] }'
PASS 'function CDATASection() { [native code] }' is 'function CDATASection() { [native code] }'
PASS 'function CSSFontFaceRule() { [native code] }' is 'function CSSFontFaceRule() { [native code] }'
PASS 'function CSSFontFaceRule() { [native code] }' is 'function CSSFontFaceRule() { [native code] }'
PASS 'function CSSImportRule() { [native code] }' is 'function CSSImportRule() { [native code] }'
PASS 'function CSSImportRule() { [native code] }' is 'function CSSImportRule() { [native code] }'
PASS 'function CSSMediaRule() { [native code] }' is 'function CSSMediaRule() { [native code] }'
PASS 'function CSSMediaRule() { [native code] }' is 'function CSSMediaRule() { [native code] }'
PASS 'function CSSPageRule() { [native code] }' is 'function CSSPageRule() { [native code] }'
PASS 'function CSSPageRule() { [native code] }' is 'function CSSPageRule() { [native code] }'
PASS 'function CSSRule() { [native code] }' is 'function CSSRule() { [native code] }'
PASS 'function CSSRule() { [native code] }' is 'function CSSRule() { [native code] }'
PASS 'function CSSRuleList() { [native code] }' is 'function CSSRuleList() { [native code] }'
PASS 'function CSSRuleList() { [native code] }' is 'function CSSRuleList() { [native code] }'
PASS 'function CSSStyleDeclaration() { [native code] }' is 'function CSSStyleDeclaration() { [native code] }'
PASS 'function CSSStyleDeclaration() { [native code] }' is 'function CSSStyleDeclaration() { [native code] }'
PASS 'function CSSStyleRule() { [native code] }' is 'function CSSStyleRule() { [native code] }'
PASS 'function CSSStyleRule() { [native code] }' is 'function CSSStyleRule() { [native code] }'
PASS 'function CSSStyleSheet() { [native code] }' is 'function CSSStyleSheet() { [native code] }'
PASS 'function CSSStyleSheet() { [native code] }' is 'function CSSStyleSheet() { [native code] }'
PASS 'function CharacterData() { [native code] }' is 'function CharacterData() { [native code] }'
PASS 'function CharacterData() { [native code] }' is 'function CharacterData() { [native code] }'
PASS 'function Comment() { [native code] }' is 'function Comment() { [native code] }'
PASS 'function Comment() { [native code] }' is 'function Comment() { [native code] }'
PASS 'function DOMException() { [native code] }' is 'function DOMException() { [native code] }'
PASS 'function DOMException() { [native code] }' is 'function DOMException() { [native code] }'
PASS 'function DOMImplementation() { [native code] }' is 'function DOMImplementation() { [native code] }'
PASS 'function DOMImplementation() { [native code] }' is 'function DOMImplementation() { [native code] }'
PASS 'function DOMParser() { [native code] }' is 'function DOMParser() { [native code] }'
PASS 'function DOMParser() { [native code] }' is 'function DOMParser() { [native code] }'
PASS 'function Document() { [native code] }' is 'function Document() { [native code] }'
PASS 'function Document() { [native code] }' is 'function Document() { [native code] }'
PASS 'function DocumentFragment() { [native code] }' is 'function DocumentFragment() { [native code] }'
PASS 'function DocumentFragment() { [native code] }' is 'function DocumentFragment() { [native code] }'
PASS 'function DocumentType() { [native code] }' is 'function DocumentType() { [native code] }'
PASS 'function DocumentType() { [native code] }' is 'function DocumentType() { [native code] }'
PASS 'function Element() { [native code] }' is 'function Element() { [native code] }'
PASS 'function Element() { [native code] }' is 'function Element() { [native code] }'
PASS 'function Event() { [native code] }' is 'function Event() { [native code] }'
PASS 'function Event() { [native code] }' is 'function Event() { [native code] }'
PASS 'function HTMLAnchorElement() { [native code] }' is 'function HTMLAnchorElement() { [native code] }'
PASS 'function HTMLAnchorElement() { [native code] }' is 'function HTMLAnchorElement() { [native code] }'
PASS 'function HTMLAreaElement() { [native code] }' is 'function HTMLAreaElement() { [native code] }'
PASS 'function HTMLAreaElement() { [native code] }' is 'function HTMLAreaElement() { [native code] }'
PASS 'function HTMLBRElement() { [native code] }' is 'function HTMLBRElement() { [native code] }'
PASS 'function HTMLBRElement() { [native code] }' is 'function HTMLBRElement() { [native code] }'
PASS 'function HTMLBaseElement() { [native code] }' is 'function HTMLBaseElement() { [native code] }'
PASS 'function HTMLBaseElement() { [native code] }' is 'function HTMLBaseElement() { [native code] }'
PASS 'function HTMLBodyElement() { [native code] }' is 'function HTMLBodyElement() { [native code] }'
PASS 'function HTMLBodyElement() { [native code] }' is 'function HTMLBodyElement() { [native code] }'
PASS 'function HTMLButtonElement() { [native code] }' is 'function HTMLButtonElement() { [native code] }'
PASS 'function HTMLButtonElement() { [native code] }' is 'function HTMLButtonElement() { [native code] }'
PASS 'function HTMLCanvasElement() { [native code] }' is 'function HTMLCanvasElement() { [native code] }'
PASS 'function HTMLCanvasElement() { [native code] }' is 'function HTMLCanvasElement() { [native code] }'
PASS 'function HTMLDListElement() { [native code] }' is 'function HTMLDListElement() { [native code] }'
PASS 'function HTMLDListElement() { [native code] }' is 'function HTMLDListElement() { [native code] }'
PASS 'function HTMLDataListElement() { [native code] }' is 'function HTMLDataListElement() { [native code] }'
PASS 'function HTMLDataListElement() { [native code] }' is 'function HTMLDataListElement() { [native code] }'
PASS 'function HTMLDirectoryElement() { [native code] }' is 'function HTMLDirectoryElement() { [native code] }'
PASS 'function HTMLDirectoryElement() { [native code] }' is 'function HTMLDirectoryElement() { [native code] }'
PASS 'function HTMLDivElement() { [native code] }' is 'function HTMLDivElement() { [native code] }'
PASS 'function HTMLDivElement() { [native code] }' is 'function HTMLDivElement() { [native code] }'
PASS 'function HTMLDocument() { [native code] }' is 'function HTMLDocument() { [native code] }'
PASS 'function HTMLDocument() { [native code] }' is 'function HTMLDocument() { [native code] }'
PASS 'function HTMLElement() { [native code] }' is 'function HTMLElement() { [native code] }'
PASS 'function HTMLElement() { [native code] }' is 'function HTMLElement() { [native code] }'
PASS 'function HTMLEmbedElement() { [native code] }' is 'function HTMLEmbedElement() { [native code] }'
PASS 'function HTMLEmbedElement() { [native code] }' is 'function HTMLEmbedElement() { [native code] }'
PASS 'function HTMLFieldSetElement() { [native code] }' is 'function HTMLFieldSetElement() { [native code] }'
PASS 'function HTMLFieldSetElement() { [native code] }' is 'function HTMLFieldSetElement() { [native code] }'
PASS 'function HTMLFontElement() { [native code] }' is 'function HTMLFontElement() { [native code] }'
PASS 'function HTMLFontElement() { [native code] }' is 'function HTMLFontElement() { [native code] }'
PASS 'function HTMLFormElement() { [native code] }' is 'function HTMLFormElement() { [native code] }'
PASS 'function HTMLFormElement() { [native code] }' is 'function HTMLFormElement() { [native code] }'
PASS 'function HTMLFrameElement() { [native code] }' is 'function HTMLFrameElement() { [native code] }'
PASS 'function HTMLFrameElement() { [native code] }' is 'function HTMLFrameElement() { [native code] }'
PASS 'function HTMLFrameSetElement() { [native code] }' is 'function HTMLFrameSetElement() { [native code] }'
PASS 'function HTMLFrameSetElement() { [native code] }' is 'function HTMLFrameSetElement() { [native code] }'
PASS 'function HTMLHRElement() { [native code] }' is 'function HTMLHRElement() { [native code] }'
PASS 'function HTMLHRElement() { [native code] }' is 'function HTMLHRElement() { [native code] }'
PASS 'function HTMLHeadElement() { [native code] }' is 'function HTMLHeadElement() { [native code] }'
PASS 'function HTMLHeadElement() { [native code] }' is 'function HTMLHeadElement() { [native code] }'
PASS 'function HTMLHeadingElement() { [native code] }' is 'function HTMLHeadingElement() { [native code] }'
PASS 'function HTMLHeadingElement() { [native code] }' is 'function HTMLHeadingElement() { [native code] }'
PASS 'function HTMLHtmlElement() { [native code] }' is 'function HTMLHtmlElement() { [native code] }'
PASS 'function HTMLHtmlElement() { [native code] }' is 'function HTMLHtmlElement() { [native code] }'
PASS 'function HTMLIFrameElement() { [native code] }' is 'function HTMLIFrameElement() { [native code] }'
PASS 'function HTMLIFrameElement() { [native code] }' is 'function HTMLIFrameElement() { [native code] }'
PASS 'function HTMLImageElement() { [native code] }' is 'function HTMLImageElement() { [native code] }'
PASS 'function HTMLImageElement() { [native code] }' is 'function HTMLImageElement() { [native code] }'
PASS 'function HTMLInputElement() { [native code] }' is 'function HTMLInputElement() { [native code] }'
PASS 'function HTMLInputElement() { [native code] }' is 'function HTMLInputElement() { [native code] }'
PASS 'function HTMLLIElement() { [native code] }' is 'function HTMLLIElement() { [native code] }'
PASS 'function HTMLLIElement() { [native code] }' is 'function HTMLLIElement() { [native code] }'
PASS 'function HTMLLabelElement() { [native code] }' is 'function HTMLLabelElement() { [native code] }'
PASS 'function HTMLLabelElement() { [native code] }' is 'function HTMLLabelElement() { [native code] }'
PASS 'function HTMLLegendElement() { [native code] }' is 'function HTMLLegendElement() { [native code] }'
PASS 'function HTMLLegendElement() { [native code] }' is 'function HTMLLegendElement() { [native code] }'
PASS 'function HTMLLinkElement() { [native code] }' is 'function HTMLLinkElement() { [native code] }'
PASS 'function HTMLLinkElement() { [native code] }' is 'function HTMLLinkElement() { [native code] }'
PASS 'function HTMLMapElement() { [native code] }' is 'function HTMLMapElement() { [native code] }'
PASS 'function HTMLMapElement() { [native code] }' is 'function HTMLMapElement() { [native code] }'
PASS 'function HTMLMarqueeElement() { [native code] }' is 'function HTMLMarqueeElement() { [native code] }'
PASS 'function HTMLMarqueeElement() { [native code] }' is 'function HTMLMarqueeElement() { [native code] }'
PASS 'function HTMLMenuElement() { [native code] }' is 'function HTMLMenuElement() { [native code] }'
PASS 'function HTMLMenuElement() { [native code] }' is 'function HTMLMenuElement() { [native code] }'
PASS 'function HTMLMetaElement() { [native code] }' is 'function HTMLMetaElement() { [native code] }'
PASS 'function HTMLMetaElement() { [native code] }' is 'function HTMLMetaElement() { [native code] }'
PASS 'function HTMLModElement() { [native code] }' is 'function HTMLModElement() { [native code] }'
PASS 'function HTMLModElement() { [native code] }' is 'function HTMLModElement() { [native code] }'
PASS 'function HTMLOListElement() { [native code] }' is 'function HTMLOListElement() { [native code] }'
PASS 'function HTMLOListElement() { [native code] }' is 'function HTMLOListElement() { [native code] }'
PASS 'function HTMLObjectElement() { [native code] }' is 'function HTMLObjectElement() { [native code] }'
PASS 'function HTMLObjectElement() { [native code] }' is 'function HTMLObjectElement() { [native code] }'
PASS 'function HTMLOptGroupElement() { [native code] }' is 'function HTMLOptGroupElement() { [native code] }'
PASS 'function HTMLOptGroupElement() { [native code] }' is 'function HTMLOptGroupElement() { [native code] }'
PASS 'function HTMLOptionElement() { [native code] }' is 'function HTMLOptionElement() { [native code] }'
PASS 'function HTMLOptionElement() { [native code] }' is 'function HTMLOptionElement() { [native code] }'
PASS 'function HTMLParagraphElement() { [native code] }' is 'function HTMLParagraphElement() { [native code] }'
PASS 'function HTMLParagraphElement() { [native code] }' is 'function HTMLParagraphElement() { [native code] }'
PASS 'function HTMLParamElement() { [native code] }' is 'function HTMLParamElement() { [native code] }'
PASS 'function HTMLParamElement() { [native code] }' is 'function HTMLParamElement() { [native code] }'
PASS 'function HTMLPreElement() { [native code] }' is 'function HTMLPreElement() { [native code] }'
PASS 'function HTMLPreElement() { [native code] }' is 'function HTMLPreElement() { [native code] }'
PASS 'function HTMLQuoteElement() { [native code] }' is 'function HTMLQuoteElement() { [native code] }'
PASS 'function HTMLQuoteElement() { [native code] }' is 'function HTMLQuoteElement() { [native code] }'
PASS 'function HTMLScriptElement() { [native code] }' is 'function HTMLScriptElement() { [native code] }'
PASS 'function HTMLScriptElement() { [native code] }' is 'function HTMLScriptElement() { [native code] }'
PASS 'function HTMLSelectElement() { [native code] }' is 'function HTMLSelectElement() { [native code] }'
PASS 'function HTMLSelectElement() { [native code] }' is 'function HTMLSelectElement() { [native code] }'
PASS 'function HTMLStyleElement() { [native code] }' is 'function HTMLStyleElement() { [native code] }'
PASS 'function HTMLStyleElement() { [native code] }' is 'function HTMLStyleElement() { [native code] }'
PASS 'function HTMLTableCaptionElement() { [native code] }' is 'function HTMLTableCaptionElement() { [native code] }'
PASS 'function HTMLTableCaptionElement() { [native code] }' is 'function HTMLTableCaptionElement() { [native code] }'
PASS 'function HTMLTableCellElement() { [native code] }' is 'function HTMLTableCellElement() { [native code] }'
PASS 'function HTMLTableCellElement() { [native code] }' is 'function HTMLTableCellElement() { [native code] }'
PASS 'function HTMLTableColElement() { [native code] }' is 'function HTMLTableColElement() { [native code] }'
PASS 'function HTMLTableColElement() { [native code] }' is 'function HTMLTableColElement() { [native code] }'
PASS 'function HTMLTableElement() { [native code] }' is 'function HTMLTableElement() { [native code] }'
PASS 'function HTMLTableElement() { [native code] }' is 'function HTMLTableElement() { [native code] }'
PASS 'function HTMLTableRowElement() { [native code] }' is 'function HTMLTableRowElement() { [native code] }'
PASS 'function HTMLTableRowElement() { [native code] }' is 'function HTMLTableRowElement() { [native code] }'
PASS 'function HTMLTableSectionElement() { [native code] }' is 'function HTMLTableSectionElement() { [native code] }'
PASS 'function HTMLTableSectionElement() { [native code] }' is 'function HTMLTableSectionElement() { [native code] }'
PASS 'function HTMLTextAreaElement() { [native code] }' is 'function HTMLTextAreaElement() { [native code] }'
PASS 'function HTMLTextAreaElement() { [native code] }' is 'function HTMLTextAreaElement() { [native code] }'
PASS 'function HTMLTitleElement() { [native code] }' is 'function HTMLTitleElement() { [native code] }'
PASS 'function HTMLTitleElement() { [native code] }' is 'function HTMLTitleElement() { [native code] }'
PASS 'function HTMLUListElement() { [native code] }' is 'function HTMLUListElement() { [native code] }'
PASS 'function HTMLUListElement() { [native code] }' is 'function HTMLUListElement() { [native code] }'
PASS 'function KeyboardEvent() { [native code] }' is 'function KeyboardEvent() { [native code] }'
PASS 'function KeyboardEvent() { [native code] }' is 'function KeyboardEvent() { [native code] }'
PASS 'function MediaList() { [native code] }' is 'function MediaList() { [native code] }'
PASS 'function MediaList() { [native code] }' is 'function MediaList() { [native code] }'
PASS 'function MouseEvent() { [native code] }' is 'function MouseEvent() { [native code] }'
PASS 'function MouseEvent() { [native code] }' is 'function MouseEvent() { [native code] }'
PASS 'function MutationEvent() { [native code] }' is 'function MutationEvent() { [native code] }'
PASS 'function MutationEvent() { [native code] }' is 'function MutationEvent() { [native code] }'
PASS 'function NamedNodeMap() { [native code] }' is 'function NamedNodeMap() { [native code] }'
PASS 'function NamedNodeMap() { [native code] }' is 'function NamedNodeMap() { [native code] }'
PASS 'function Node() { [native code] }' is 'function Node() { [native code] }'
PASS 'function Node() { [native code] }' is 'function Node() { [native code] }'
PASS 'function NodeFilter() { [native code] }' is 'function NodeFilter() { [native code] }'
PASS 'function NodeFilter() { [native code] }' is 'function NodeFilter() { [native code] }'
PASS 'function NodeList() { [native code] }' is 'function NodeList() { [native code] }'
PASS 'function NodeList() { [native code] }' is 'function NodeList() { [native code] }'
PASS 'function ProcessingInstruction() { [native code] }' is 'function ProcessingInstruction() { [native code] }'
PASS 'function ProcessingInstruction() { [native code] }' is 'function ProcessingInstruction() { [native code] }'
PASS 'function ProgressEvent() { [native code] }' is 'function ProgressEvent() { [native code] }'
PASS 'function ProgressEvent() { [native code] }' is 'function ProgressEvent() { [native code] }'
PASS 'function Range() { [native code] }' is 'function Range() { [native code] }'
PASS 'function Range() { [native code] }' is 'function Range() { [native code] }'
PASS 'function SVGAngle() { [native code] }' is 'function SVGAngle() { [native code] }'
PASS 'function SVGAngle() { [native code] }' is 'function SVGAngle() { [native code] }'
PASS 'function SVGGradientElement() { [native code] }' is 'function SVGGradientElement() { [native code] }'
PASS 'function SVGGradientElement() { [native code] }' is 'function SVGGradientElement() { [native code] }'
PASS 'function SVGLength() { [native code] }' is 'function SVGLength() { [native code] }'
PASS 'function SVGLength() { [native code] }' is 'function SVGLength() { [native code] }'
PASS 'function SVGMarkerElement() { [native code] }' is 'function SVGMarkerElement() { [native code] }'
PASS 'function SVGMarkerElement() { [native code] }' is 'function SVGMarkerElement() { [native code] }'
PASS 'function SVGPreserveAspectRatio() { [native code] }' is 'function SVGPreserveAspectRatio() { [native code] }'
PASS 'function SVGPreserveAspectRatio() { [native code] }' is 'function SVGPreserveAspectRatio() { [native code] }'
PASS 'function SVGTextContentElement() { [native code] }' is 'function SVGTextContentElement() { [native code] }'
PASS 'function SVGTextContentElement() { [native code] }' is 'function SVGTextContentElement() { [native code] }'
PASS 'function SVGTextPathElement() { [native code] }' is 'function SVGTextPathElement() { [native code] }'
PASS 'function SVGTextPathElement() { [native code] }' is 'function SVGTextPathElement() { [native code] }'
PASS 'function SVGTransform() { [native code] }' is 'function SVGTransform() { [native code] }'
PASS 'function SVGTransform() { [native code] }' is 'function SVGTransform() { [native code] }'
PASS 'function SVGUnitTypes() { [native code] }' is 'function SVGUnitTypes() { [native code] }'
PASS 'function SVGUnitTypes() { [native code] }' is 'function SVGUnitTypes() { [native code] }'
PASS 'function StyleSheet() { [native code] }' is 'function StyleSheet() { [native code] }'
PASS 'function StyleSheet() { [native code] }' is 'function StyleSheet() { [native code] }'
PASS 'function StyleSheetList() { [native code] }' is 'function StyleSheetList() { [native code] }'
PASS 'function StyleSheetList() { [native code] }' is 'function StyleSheetList() { [native code] }'
PASS 'function Text() { [native code] }' is 'function Text() { [native code] }'
PASS 'function Text() { [native code] }' is 'function Text() { [native code] }'
PASS 'function TextEvent() { [native code] }' is 'function TextEvent() { [native code] }'
PASS 'function TextEvent() { [native code] }' is 'function TextEvent() { [native code] }'
PASS 'function UIEvent() { [native code] }' is 'function UIEvent() { [native code] }'
PASS 'function UIEvent() { [native code] }' is 'function UIEvent() { [native code] }'
PASS 'function WheelEvent() { [native code] }' is 'function WheelEvent() { [native code] }'
PASS 'function WheelEvent() { [native code] }' is 'function WheelEvent() { [native code] }'
PASS 'function XMLDocument() { [native code] }' is 'function XMLDocument() { [native code] }'
PASS 'function XMLDocument() { [native code] }' is 'function XMLDocument() { [native code] }'
PASS 'function XMLHttpRequest() { [native code] }' is 'function XMLHttpRequest() { [native code] }'
PASS 'function XMLHttpRequest() { [native code] }' is 'function XMLHttpRequest() { [native code] }'
PASS 'function XMLSerializer() { [native code] }' is 'function XMLSerializer() { [native code] }'
PASS 'function XMLSerializer() { [native code] }' is 'function XMLSerializer() { [native code] }'
PASS 'function XPathEvaluator() { [native code] }' is 'function XPathEvaluator() { [native code] }'
PASS 'function XPathEvaluator() { [native code] }' is 'function XPathEvaluator() { [native code] }'
PASS 'function XPathResult() { [native code] }' is 'function XPathResult() { [native code] }'
PASS 'function XPathResult() { [native code] }' is 'function XPathResult() { [native code] }'
PASS 'function XSLTProcessor() { [native code] }' is 'function XSLTProcessor() { [native code] }'
PASS 'function XSLTProcessor() { [native code] }' is 'function XSLTProcessor() { [native code] }'
PASS win['onload'] is null
PASS win['onload'] is null
PASS win['frames'] == '[object Window]' is true
PASS win['frames'] == '[object Window]' is true
PASS win['parent'] == '[object Window]' is true
PASS win['parent'] == '[object Window]' is true
PASS win['self'] == '[object Window]' is true
PASS win['self'] == '[object Window]' is true
PASS win['top'] == '[object Window]' is true
PASS win['top'] == '[object Window]' is true
PASS win['window'] == '[object Window]' is true
PASS win['window'] == '[object Window]' is true
PASS win['locationbar'] == '[object BarProp]' is true
PASS win['locationbar'] == '[object BarProp]' is true
PASS win['menubar'] == '[object BarProp]' is true
PASS win['menubar'] == '[object BarProp]' is true
PASS win['personalbar'] == '[object BarProp]' is true
PASS win['personalbar'] == '[object BarProp]' is true
PASS win['scrollbars'] == '[object BarProp]' is true
PASS win['scrollbars'] == '[object BarProp]' is true
PASS win['statusbar'] == '[object BarProp]' is true
PASS win['statusbar'] == '[object BarProp]' is true
PASS win['toolbar'] == '[object BarProp]' is true
PASS win['toolbar'] == '[object BarProp]' is true
PASS win['console'] == '[object Object]' is true
PASS win['history'] == '[object History]' is true
PASS win['navigator'] == '[object Navigator]' is true
PASS win['navigator'] == '[object Navigator]' is true
PASS win['screen'] == '[object Screen]' is true
PASS win['screen'] == '[object Screen]' is true
PASS win['clientInformation'] == '[object Navigator]' is true
PASS win['clientInformation'] == '[object Navigator]' is true
PASS win['document'] == '[object HTMLDocument]' is true
PASS win['document'] == '[object HTMLDocument]' is true
PASS win['location'] matched original
PASS win['location'] matched original
PASS win['onabort'] is null
PASS win['onabort'] is null
PASS win['onbeforeunload'] is null
PASS win['onbeforeunload'] is null
PASS win['onblur'] is null
PASS win['onblur'] is null
PASS win['onchange'] is null
PASS win['onchange'] is null
PASS win['onclick'] is null
PASS win['onclick'] is null
PASS win['ondblclick'] is null
PASS win['ondblclick'] is null
PASS win['onerror'] is null
PASS win['onerror'] is null
PASS win['onfocus'] is null
PASS win['onfocus'] is null
PASS win['onkeydown'] is null
PASS win['onkeydown'] is null
PASS win['onkeypress'] is null
PASS win['onkeypress'] is null
PASS win['onkeyup'] is null
PASS win['onkeyup'] is null
PASS win['onmousedown'] is null
PASS win['onmousedown'] is null
PASS win['onmousemove'] is null
PASS win['onmousemove'] is null
PASS win['onmouseout'] is null
PASS win['onmouseout'] is null
PASS win['onmouseover'] is null
PASS win['onmouseover'] is null
PASS win['onmouseup'] is null
PASS win['onmouseup'] is null
PASS win['onmousewheel'] is null
PASS win['onmousewheel'] is null
PASS win['onreset'] is null
PASS win['onreset'] is null
PASS win['onresize'] is null
PASS win['onresize'] is null
PASS win['onscroll'] is null
PASS win['onscroll'] is null
PASS win['onsearch'] is null
PASS win['onsearch'] is null
PASS win['onselect'] is null
PASS win['onselect'] is null
PASS win['onsubmit'] is null
PASS win['onsubmit'] is null
PASS win['onunload'] is null
PASS win['onunload'] is null
PASS win['opener'] is null
PASS win['opener'] is null
FAIL win['defaultStatus'] should be undefined (of type undefined). Was  (of type string).
FAIL win['defaultStatus'] should be undefined (of type undefined). Was  (of type string).
FAIL win['defaultstatus'] should be undefined (of type undefined). Was  (of type string).
FAIL win['defaultstatus'] should be undefined (of type undefined). Was  (of type string).
FAIL win['name'] should be undefined (of type undefined). Was  (of type string).
FAIL win['name'] should be undefined (of type undefined). Was  (of type string).
FAIL win['status'] should be undefined (of type undefined). Was  (of type string).
FAIL win['status'] should be undefined (of type undefined). Was  (of type string).
PASS win['devicePixelRatio'] matched original
PASS win['devicePixelRatio'] matched original
PASS win['innerHeight'] matched original
PASS win['innerHeight'] matched original
PASS win['innerWidth'] matched original
PASS win['innerWidth'] matched original
PASS win['length'] matched original
PASS win['length'] matched original
PASS win['outerHeight'] matched original
PASS win['outerHeight'] matched original
PASS win['outerWidth'] matched original
PASS win['outerWidth'] matched original
PASS win['pageXOffset'] matched original
PASS win['pageXOffset'] matched original
PASS win['pageYOffset'] matched original
PASS win['pageYOffset'] matched original
PASS win['screenLeft'] matched original
PASS win['screenLeft'] matched original
PASS win['screenTop'] matched original
PASS win['screenTop'] matched original
PASS win['screenX'] matched original
PASS win['screenX'] matched original
PASS win['screenY'] matched original
PASS win['screenY'] matched original
PASS win['scrollX'] matched original
PASS win['scrollX'] matched original
PASS win['scrollY'] matched original
PASS win['scrollY'] matched original
PASS typeof win['closed'] is 'boolean'
PASS win['closed'] is false
PASS typeof win['closed'] is 'boolean'
PASS win['closed'] is false
PASS typeof win['offscreenBuffering'] is 'boolean'
PASS win['offscreenBuffering'] is true
PASS typeof win['offscreenBuffering'] is 'boolean'
PASS win['offscreenBuffering'] is true
PASS successfullyParsed is true

TEST COMPLETE
